+++ src/main/kotlin/omok/InputExceptionHandler.kt
@@ -1,0 +1,10 @@
+package omok
+
+fun <T> retryWhileNotException(block: () -> T): T {
+    return try {
+        block()
+    } catch (e: IllegalArgumentException) {
+        println(e.message)
+        retryWhileNotException(block)
+    }
+}
